Refactor: create storage package
